+++ assets-src/js/cookie.js
@@ -1,25 +1,21 @@
-
 const $ = window.jQuery;
 
 export default class Cookie {
+  constructor() {}
 
-  constructor() {
-
-  }
-
-  static get( name, fallback = null ) {
-    var v = document.cookie.match('(^|;) ?' + name + '=([^;]*)(;|$)');
+  static get(name, fallback = null) {
+    var v = document.cookie.match("(^|;) ?" + name + "=([^;]*)(;|$)");
     return v ? v[2] : fallback;
   }
 
-  static set(name, value, days = 10000 ) {
-    var d = new Date;
-    d.setTime(d.getTime() + 24*60*60*1000*days);
+  static set(name, value, days = 10000) {
+    var d = new Date();
+    d.setTime(d.getTime() + 24 * 60 * 60 * 1000 * days);
     document.cookie = name + "=" + value + ";path=/;expires=" + d.toGMTString();
   }
 
-  static delete( name ) {
-    Cookie.set(name, '', 0);
+  static delete(name) {
+    Cookie.set(name, "", 0);
   }
 }