+++ lib/rh-admin-utils/Misc.php
@@ -292,12 +292,12 @@ class Misc extends Singleton
         $post_id = $field['value'] ?? null;
         if (empty($post_id) || !$post = get_post($post_id)) return;
 
-?>
+        ?>
         <div class="rh-post-object-edit-links">
             <a href="<?= get_edit_post_link($post_id) ?>">Edit</a>
             <a href="<?= get_permalink($post_id) ?>" target="_blank">View</a>
         </div>
-    <?php
+        <?php
     }
 
     /**
@@ -305,7 +305,7 @@ class Misc extends Singleton
      */
     public static function render_acf_post_object_styles()
     {
-    ?>
+        ?>
         <style>
             .rh-post-object-edit-links {
                 position: absolute;
@@ -322,7 +322,7 @@ class Misc extends Singleton
                 text-decoration: none;
             }
         </style>
-<?php
+        <?php
     }
 
     /**